Move everything from general CoaxialSwerve to Kraken
