+++ src/main/java/com/techhounds/houndutil/houndlib/swerve/KrakenCoaxialSwerveModule.java
@@ -26,6 +26,7 @@ import com.techhounds.houndutil.houndlog.annotations.LoggedObject;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.kinematics.SwerveModulePosition;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
+import edu.wpi.first.math.system.plant.DCMotor;
 import edu.wpi.first.math.system.plant.LinearSystemId;
 import edu.wpi.first.units.measure.Angle;
 import edu.wpi.first.units.measure.AngularAcceleration;
@@ -43,7 +44,146 @@ import edu.wpi.first.wpilibj.simulation.DCMotorSim;
  * For status signals to update, you must call SignalManager.update().
  */
 @LoggedObject
-public class KrakenCoaxialSwerveModule implements CoaxialSwerveModule {
+public class KrakenCoaxialSwerveModule {
+    /**
+     * Common constants for a coaxial swerve module. To initialize, create a
+     * `static` block in your Constants class and set the values.
+     */
+    public static class SwerveConstants {
+        /** The P gain for feedback velocity control on the drive wheel. */
+        public double DRIVE_kP;
+        /** The I gain for feedback velocity control on the drive wheel. */
+        public double DRIVE_kI;
+        /** The D gain for feedback velocity control on the drive wheel. */
+        public double DRIVE_kD;
+        /** The S gain for feedforward velocity control on the drive wheel. */
+        public double DRIVE_kS;
+        /** The V gain for feedforward velocity control on the drive wheel. */
+        public double DRIVE_kV;
+        /** The A gain for feedforward velocity control on the drive wheel. */
+        public double DRIVE_kA;
+
+        /** The P gain for feedback position control on the steer wheel. */
+        public double STEER_kP;
+        /** The I gain for feedback position control on the steer wheel. */
+        public double STEER_kI;
+        /** The D gain for feedback position control on the steer wheel. */
+        public double STEER_kD;
+        /** The S gain for feedforward position control on the steer wheel. */
+        public double STEER_kS;
+        /** The V gain for feedforward position control on the steer wheel. */
+        public double STEER_kV;
+        /** The A gain for feedforward position control on the steer wheel. */
+        public double STEER_kA;
+
+        /**
+         * The number of rotations the drive motor makes for every one rotation of the
+         * wheel.
+         */
+        public double DRIVE_GEARING;
+        /**
+         * The number of rotations the steer motor makes for every one azimuth rotation
+         * of the wheel.
+         */
+        public double STEER_GEARING;
+        /**
+         * For a typical swerve module, the azimuth turn motor also drives the wheel a
+         * nontrivial amount, which affects the accuracy of odometry and control. This
+         * ratio
+         * represents the number of rotations of the drive motor caused by a rotation of
+         * the azimuth.
+         */
+        public double COUPLING_RATIO;
+        /**
+         * Multiplier to convert drive motor rotations to meters travelled by the wheel.
+         * Typically WHEEL_CIRCUMFERENCE / DRIVE_GEARING.
+         */
+        public double DRIVE_ENCODER_ROTATIONS_TO_METERS;
+        /**
+         * Multiplier to convert steer motor rotations to the overall rotation of the
+         * wheel.
+         * Typically 2 * Math.PI / STEER_GEARING.
+         */
+        public double STEER_ENCODER_ROTATIONS_TO_RADIANS;
+        /**
+         * The circumference of the wheel, in meters. This is the radius of the wheel *
+         * 2pi. This is best determined experimentally, as different wheel types
+         * compress differently and produce different effective wheel radii. You should
+         * update this value relatively often, as wheel wear will reduce the effective
+         * wheel radius, which is used for positional control.
+         */
+        public double WHEEL_CIRCUMFERENCE;
+
+        /**
+         * The maximum velocity of the edge of the wheel in meters per second. This can
+         * be determined theoretically based on the maximum RPM of the motor and the
+         * gear ratio, but it is recommended to determine this empirically as the
+         * effective maximum velocity will be slightly lower.
+         * 
+         * <p>
+         * This is effectively the maximum velocity of the robot.
+         */
+        public double MAX_DRIVING_VELOCITY_METERS_PER_SECOND;
+        /**
+         * The maximum acceleration of the edge of the wheel in meters per second per
+         * second. This should be determined empirically.
+         * 
+         * <p>
+         * This is effectively the maximum acceleration of the robot.
+         */
+        public double MAX_DRIVING_ACCELERATION_METERS_PER_SECOND_SQUARED;
+        /**
+         * The maximum rotational velocity of the azimuth of the wheel. This can be
+         * determined theoretically based on the maximum RPM of the motor and the
+         * steering gear ratio, but it is recommended to determine this empirically as
+         * the effective maximum velocity will be slightly lower.
+         */
+        public double MAX_STEER_VELOCITY_RADIANS_PER_SECOND;
+        /**
+         * The maximum rotational acceleration of the azimuth of the wheel. This can be
+         * determined
+         */
+        public double MAX_STEER_ACCELERATION_RADIANS_PER_SECOND_SQUARED;
+
+        /**
+         * The stator current limit of the drive motor.
+         * 
+         * <p>
+         * 
+         * Since the torque output of a motor is directly proportional to stator
+         * current, this should be set to the maximum value that will not allow the
+         * wheels to slip. This number will be higher for higher-traction wheels. Do not
+         * exceed the thermal specifications of the motor.
+         */
+        public int DRIVE_CURRENT_LIMIT;
+        /**
+         * The stator current limit of the steer motor.
+         */
+        public int STEER_CURRENT_LIMIT;
+
+        /**
+         * A {@link DCMotor} representation of the driving motor. On a typical swerve
+         * module, call {@code getMotorName(1)} (i.e. {@code getKrakenX60(1)}).
+         */
+        public DCMotor DRIVE_GEARBOX_REPR;
+        /**
+         * A {@link DCMotor} representation of the steer motor. On a typical swerve
+         * module, call {@code getMotorName(1)} (i.e. {@code getKrakenX60(1)}).
+         */
+        public DCMotor STEER_GEARBOX_REPR;
+
+        /**
+         * The moment of inertia of the drive motor in kg*m/s^2. Used for simulation. If
+         * you are unsure, use 0.01.
+         */
+        public double DRIVE_MOI;
+        /**
+         * The moment of inertia of the steer motor in kg*m/s^2. Used for simulation. If
+         * you are unsure, use 0.025.
+         */
+        public double STEER_MOI;
+    }
+
     /** The motor used for driving. */
     @Log
     private TalonFX driveMotor;
@@ -188,17 +328,14 @@ public class KrakenCoaxialSwerveModule implements CoaxialSwerveModule {
                 steerPosition, steerVelocity, steerAcceleration, steerMotorVoltage);
     }
 
-    @Override
     public double getDriveMotorPosition() {
         return BaseStatusSignal.getLatencyCompensatedValue(drivePosition, driveVelocity).magnitude();
     }
 
-    @Override
     public double getDriveMotorVelocity() {
         return BaseStatusSignal.getLatencyCompensatedValue(driveVelocity, driveAcceleration).magnitude();
     }
 
-    @Override
     public double getDriveMotorVoltage() {
         return driveMotorVoltage.getValue().magnitude();
     }
@@ -212,17 +349,14 @@ public class KrakenCoaxialSwerveModule implements CoaxialSwerveModule {
         return driveMotor;
     }
 
-    @Override
     public double getSteerMotorPosition() {
         return BaseStatusSignal.getLatencyCompensatedValue(steerPosition, steerVelocity).magnitude();
     }
 
-    @Override
     public double getSteerMotorVelocity() {
         return BaseStatusSignal.getLatencyCompensatedValue(steerVelocity, steerAcceleration).magnitude();
     }
 
-    @Override
     public double getSteerMotorVoltage() {
         return steerMotorVoltage.getValue().magnitude();
     }
@@ -245,12 +379,10 @@ public class KrakenCoaxialSwerveModule implements CoaxialSwerveModule {
         return new BaseStatusSignal[] { drivePosition, driveVelocity, steerPosition, steerVelocity };
     }
 
-    @Override
     public Rotation2d getWheelAngle() {
         return Rotation2d.fromRotations(getSteerMotorPosition());
     }
 
-    @Override
     @Log
     public SwerveModulePosition getPosition() {
         return new SwerveModulePosition(
@@ -265,13 +397,11 @@ public class KrakenCoaxialSwerveModule implements CoaxialSwerveModule {
                 getWheelAngle());
     }
 
-    @Override
     @Log
     public SwerveModuleState getState() {
         return new SwerveModuleState(getDriveMotorVelocity() * SWERVE_CONSTANTS.WHEEL_CIRCUMFERENCE, getWheelAngle());
     }
 
-    @Override
     public void setMotorHoldMode(MotorHoldMode motorHoldMode) {
         MotorOutputConfigs driveConfigs = new MotorOutputConfigs();
         driveMotor.getConfigurator().refresh(driveConfigs);
@@ -286,7 +416,6 @@ public class KrakenCoaxialSwerveModule implements CoaxialSwerveModule {
         steerMotor.getConfigurator().apply(steerConfigs);
     }
 
-    @Override
     public void setDriveCurrentLimit(int currentLimit) {
         CurrentLimitsConfigs currentConfigs = new CurrentLimitsConfigs();
         driveMotor.getConfigurator().refresh(currentConfigs);
@@ -295,7 +424,6 @@ public class KrakenCoaxialSwerveModule implements CoaxialSwerveModule {
         driveMotor.getConfigurator().apply(currentConfigs);
     }
 
-    @Override
     public void stop() {
         driveMotor.setControl(driveVoltageRequest.withOutput(0));
         steerMotor.set(0);
@@ -340,12 +468,10 @@ public class KrakenCoaxialSwerveModule implements CoaxialSwerveModule {
         }
     }
 
-    @Override
     public void setState(SwerveModuleState state) {
         setStateInternal(state, true);
     }
 
-    @Override
     public void setStateClosedLoop(SwerveModuleState state) {
         setStateInternal(state, false);
     }
